fix classOrObjectMethod and cleanup
So my last PR made it transpile class and object methods fine, but they transpiled to the other way, which we're looking for just plain converting from async to generator. 

This PR fixes that issue, and also removes the `callId` so that it doesn't put the extra stuff at the top for the other method.